% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_dissim.R
\name{ts_dissim}
\alias{ts_dissim}
\title{Dissimilarity Index}
\usage{
ts_dissim(.data, popA, popB, dissim, return = c("index", "tibble"))
}
\arguments{
\item{.data}{A tbl}

\item{popA}{Column name of group A population count per feature}

\item{popB}{Column name of group B population count per feature}

\item{return}{Specifies output. When \code{index}, a single value will be returned.
When \code{tibble}, a table will be returned with dissimilarity values per feature
appended as a column.}
}
\value{
Based on \code{return}, either a single index of dissimilarity value or
    a tibble with dissimilarity values per feature.
}
\description{
Calculates an index of dissimilarity value for a given pair of groups.
}
